% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\name{era}
\alias{era}
\title{Create an era object}
\usage{
era(
  abbreviation,
  epoch = NULL,
  name = abbreviation,
  unit = c("calendar years", "radiocarbon years"),
  scale = 1,
  direction = c("backwards", "forwards")
)
}
\arguments{
\item{abbreviation}{Character. If only one argument is given to \code{era()}, the
abbreviation of a standard era defined in the package.
Otherwise, the abbreviation of a user-specified era defined by the following
arguments.}

\item{epoch}{Integer. Epoch year from which years are counted (in the Common
Era).}

\item{name}{Character. Full name of the era. Defaults to the value of
\code{abbreviation}.}

\item{unit}{Character. Default: \code{"calendar years"}.}

\item{scale}{Integer. Number of years represented by one unit, e.g. \code{1000}
for ka. Default: 1.}

\item{direction}{Are years counted \code{"backwards"} (the default) or \code{"forwards"}
from \code{epoch}?}
}
\value{
An object of class \code{era}.
}
\description{
The \code{era} class defines the time scale associated with a vector of years
(see \code{\link[=yr]{yr()}}). \code{era()} returns an \code{era} object, either by looking up
\code{abbreviation} in the standard definitions included in the package or, if
more than one argument is given, constructing a new definition with the
given parameters.
}
\examples{
era("cal BP")

era("T.A.", epoch = -9021, name = "Third Age", direction = "forwards")
}
