% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\name{era_attributes}
\alias{era_attributes}
\alias{era_abbreviation}
\alias{era_abbr}
\alias{era_epoch}
\alias{era_name}
\alias{era_unit}
\alias{era_scale}
\alias{era_direction}
\title{Get attributes of an era}
\usage{
era_abbreviation(x)

era_abbr(x)

era_epoch(x)

era_name(x)

era_unit(x)

era_scale(x)

era_direction(x)
}
\arguments{
\item{x}{An \code{era} object.}
}
\value{
Value of the attribute.
}
\description{
Extracts a specific attribute from an era object.
}
\details{
The available attributes are:
\itemize{
\item \strong{abbreviation} (abbr) – abbreviated name of the era, e.g. "cal BP"
\item \strong{epoch} – year of origin of the era, e.g. 1950 for years Before Present
\item \strong{name} – full name of the era, e.g. "calendar years Before Present"
\item \strong{unit} – unit of years used, e.g. "calendar years", "radiocarbon years"
\item \strong{scale} – multiple of years used, e.g. 1000 for ka/kiloannum
\item \strong{direction} – whether years are counted "backwards" or "forwards" from the epoch
}
}
\examples{
x <- era("cal BP")
era_name(x)
}
