% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\name{is_era}
\alias{is_era}
\title{Is this an \code{era} object?}
\usage{
is_era(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Tests whether an object is an \code{era}; a calendar era definition constructed
by \code{\link[=era]{era()}}.
}
\seealso{
Other era helper functions: 
\code{\link{era_parameters}},
\code{\link{is_yr}()}
}
\concept{era helper functions}
