% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era_year.R
\name{is_era_year}
\alias{is_era_year}
\title{Validation functions for \code{era_year} objects}
\usage{
is_era_year(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Tests whether an object is of class \code{era_year} (constructed by \code{\link[=era_year]{era_year()}}).
}
\examples{
is_era_year(era_year("Julian", 365.25))
}
\seealso{
Other era helper functions: 
\code{\link{era_parameters}},
\code{\link{era_year_parameters}},
\code{\link{era_year}()},
\code{\link{era}()},
\code{\link{is_era}()},
\code{\link{is_yr}()},
\code{\link{this_year}()}
}
\concept{era helper functions}
