% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr.R
\name{is_yr}
\alias{is_yr}
\title{Is this a \code{yr} vector?}
\usage{
is_yr(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Tests whether an object is a \code{yr}; a vector of years with era.
}
\examples{
x <- yr(5000:5050, era("cal BP"))
is_yr(x)

#x <- as.integer(x)
#is_yr(x)
}
\seealso{
Other era helper functions: 
\code{\link{era_parameters}},
\code{\link{is_era}()}
}
\concept{era helper functions}
