% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr.R
\name{yr}
\alias{yr}
\title{Years with era}
\usage{
yr(x = numeric(), era)
}
\arguments{
\item{x}{A numeric vector of years.}

\item{era}{Object describing time scale. Either:
\itemize{
\item A string matching one of the standard era labels defined in \code{\link[=eras]{eras()}}
\item An \code{era} object constructed with \code{\link[=era]{era()}}
}}
}
\value{
A \code{yr} (\code{era_yr}) object.
}
\description{
Represents years with an associated era or time scale.
}
\examples{
# The R Age
yr(1993:2020, "CE")

# A bad movie
yr(10000, "BC")
}
