% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{yr_transform}
\alias{yr_transform}
\title{Convert years from one era to another}
\usage{
yr_transform(x, era = yr_era(x))
}
\arguments{
\item{x}{A \code{yr} object to be transformed.}

\item{era}{Target \code{era} object, see \code{\link[=era]{era()}}.}
}
\value{
A \code{yr} object in the era specified by \code{era}.
}
\description{
Transform a vector of years from one era to another. Not all transformations
are possible (see details).
}
\details{
Era transformations currently supported:
\itemize{
\item Between scales
\item Between epochs
}

Not supported:
\itemize{
\item Between directions (to be implemented).
\item units: \verb{calendar years} ↔ \verb{radiocarbon years}. This requires \href{https://en.wikipedia.org/wiki/Radiocarbon_calibration}{calibration}
or un-calibration, which is outside the scope of this package. Functions
for radiocarbon calibration can be found in \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} and
\code{\link[Bchron:BchronCalibrate]{Bchron::BchronCalibrate()}}.
}
}
\examples{
x <- yr(10010:10001, "cal BP")
yr_transform(x, era("BCE"))
}
